% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ColorDataSets_Classic}
\alias{ColorDataSets_Classic}
\title{获取《小林家的龙女仆》经典配色数据}
\usage{
ColorDataSets_Classic()
}
\value{
一个 \code{data.frame}，包含两列：
  \describe{
    \item{\code{name}}{角色罗马音名称，因子类型，共 12 个水平，每个水平重复 5 次。}
    \item{\code{color}}{对应的十六进制颜色码，字符型。}
  }
}
\description{
返回一个包含所有登场角色经典服装配色信息的数据框。
每位角色对应 5 种颜色，来源于萌娘百科的角色形象设定。
}
\examples{
head(ColorDataSets_Classic())
}
