% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_dragon_scarf}
\alias{icon_dragon_scarf}
\title{绘制小林送给托尔的围巾（龙图案）}
\usage{
icon_dragon_scarf()
}
\value{
返回一个包含该围巾龙图案的 ggplot 对象。
}
\description{
绘制《小林家的龙女仆》中小林送给托尔的围巾上的龙图案，红色背景上有一条绿色龙纹。
}
\examples{
icon_dragon_scarf()
}
