% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_rabbit_Kobayashi}
\alias{icon_rabbit_Kobayashi}
\title{绘制小林款兔子图标}
\usage{
icon_rabbit_Kobayashi()
}
\value{
返回一个包含小林款兔子图案的 `ggplot` 对象。
}
\description{
绘制小林同款马克杯上的兔子图案（印度红兔子，金色背景）。
}
\examples{
icon_rabbit_Kobayashi()
}
