% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_rabbit_layer}
\alias{icon_rabbit_layer}
\title{绘制小林家的龙女仆同款兔子图标（ggplot图层）}
\usage{
icon_rabbit_layer(
  fill,
  linecolor = NA,
  background,
  icon_size = 1,
  x0 = 0,
  y0 = 0,
  angle
)
}
\arguments{
\item{fill}{兔子主体部分的填充颜色 (例如: 'gold', 'indianred', '#FFFFFF')。}

\item{linecolor}{图形边框的颜色。如果不需要边框，可以使用 `NA`。}

\item{background}{背景颜色。}

\item{icon_size}{图标的整体缩放比例，默认为 1。}

\item{x0}{图标中心的 X 轴坐标，默认为 0。}

\item{y0}{图标中心的 Y 轴坐标，默认为 0。}

\item{angle}{图标的旋转角度。}
}
\value{
返回一个包含该兔子图案的 `ggplot` 图层。
}
\description{
这是一个底层函数，用于通过 `ggplot2` 和 `ggforce` 绘制《小林家的龙女仆》中，小林和托尔同款马克杯上的兔子图案。
}
\examples{
# 绘制多个自定义兔子
ggplot2::ggplot()+icon_rabbit_layer(fill='lightskyblue',linecolor=NA,background=NA,icon_size=1,x0=(-2),y0=(0),angle=0)+icon_rabbit_layer(fill='darkgreen',linecolor=NA,background=NA,icon_size=2,x0=(5),y0=(0),angle=45)
}
