% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon_rabbit_Tohru}
\alias{icon_rabbit_Tohru}
\title{绘制托尔款兔子图标}
\usage{
icon_rabbit_Tohru()
}
\value{
返回一个包含托尔款兔子图案的 `ggplot` 对象。
}
\description{
绘制托尔同款马克杯上的兔子图案（金色兔子，印度红背景）。
}
\examples{
icon_rabbit_Tohru()
}
