% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{Show_Maid_colors}
\alias{Show_Maid_colors}
\title{绘制《小林家的龙女仆》角色主题色热图}
\usage{
Show_Maid_colors(lang = c("Chinese", "Japanese", "Romaji"))
}
\arguments{
\item{lang}{字符串，指定显示的语言。可选值为 \code{"Chinese"} (中文),
\code{"Japanese"} (日文), 或 \code{"Romaji"} (罗马音)。默认为 \code{"Chinese"}。}
}
\value{
返回一个 \code{ggplot} 对象，显示角色主题色的热图。
}
\description{
根据输入指定的语言（中文、日文或罗马音），生成一个展示各个角色经典配色（Hex色值）的热图。
颜色将按照调色板顺序从左到右排列，左侧 Y 轴为对应语言的角色名称。
}
\examples{
# 绘制中文角色名热图
Show_Maid_colors(lang = "Chinese")

# 绘制日文角色名热图
Show_Maid_colors(lang = "Japanese")
}
